+++ src/formatters/procedureParameter/ProcedureParameterSettings.ts
@@ -3,6 +3,6 @@ import { ASettings } from "../ASettings";
 export class ProcedureParameterSettings extends ASettings {
     // procedure parameter settings
     public procedureParameterFormatting() {
-        return !!this.configurationManager.get("procedureParamaterFormatting");
+        return !!this.configurationManager.get("procedureParameterFormatting");
     }
 }