+++ src/formatters/using/UsingSettings.ts
@@ -5,4 +5,8 @@ export class UsingSettings extends ASettings {
     public usingFormatting() {
         return !!this.configurationManager.get("usingFormatting");
     }
+
+    public usingFormattingFromPropath() {
+        return this.configurationManager.get("usingFormattingFromPropath");
+    }
 }