+++ src/test-symbol/suite/symbol.test.ts
@@ -1,14 +1,14 @@
-import * as fs from "fs";
+import * as fs from "node:fs";
 import * as vscode from "vscode";
 import { enableFormatterDecorators } from "../../formatterFramework/enableFormatterDecorators";
 import {
     setupParserHelper,
-    stabilityTestCases,
+    getStabilityTestCases,
     getTestRunDir,
     runGenericTest,
-    TestConfig,
     logKnownFailures,
 } from "../../utils/suitesUtils";
+import { ISuiteConfig } from "../../utils/ISuiteConfig";
 import { AblParserHelper } from "../../parser/AblParserHelper";
 
 let parserHelper: AblParserHelper;
@@ -16,12 +16,15 @@ let parserHelper: AblParserHelper;
 suite("Symbol Stability Test Suite", () => {
     suiteSetup(async () => {
         console.log("Symbol Test Suite setup");
+
         const symbolTestRunDir = getTestRunDir("symbolTests");
         fs.mkdirSync(symbolTestRunDir, { recursive: true });
+
         parserHelper = await setupParserHelper();
+
         console.log(
             "Symbol StabilityTests: ",
-            stabilityTestCases.length,
+            getStabilityTestCases().length,
             "test cases"
         );
 
@@ -37,35 +40,26 @@ suite("Symbol Stability Test Suite", () => {
         vscode.window.showInformationMessage("Symbol tests done!");
     });
 
-    stabilityTestCases.forEach((cases) => {
-        test(`Symbol test: ${cases}`, async () => {
-            await symbolTest(cases, parserHelper);
+    for (const cases of getStabilityTestCases()) {
+        test(`Symbol test: ${cases}`, () => {
+            symbolTest(cases, parserHelper);
         }).timeout(20000);
-    });
+    }
 });
 
-async function symbolTest(
-    name: string,
-    parserHelper: AblParserHelper
-): Promise<void> {
+function symbolTest(name: string, parserHelper: AblParserHelper): void {
     enableFormatterDecorators();
 
-    const config: TestConfig<{ tree: any; text: string }> = {
+    const config: ISuiteConfig<number> = {
         testType: "symbol",
         knownFailuresFile: "_symbol_failures.txt",
         resultFailuresFile: "_symbol_failures.txt",
-        processBeforeText: async (text: string) => ({
-            tree: null,
-            text: countActualSymbols(text).toString(),
-        }),
-        processAfterText: async (text: string) => ({
-            tree: null,
-            text: countActualSymbols(text).toString(),
-        }),
-        compareResults: async (before, after) => before.text !== after.text,
+        processBeforeText: (text: string) => countActualSymbols(text),
+        processAfterText: (text: string) => countActualSymbols(text),
+        compareResults: (before: number, after: number) => before !== after,
     };
 
-    await runGenericTest(name, parserHelper, config);
+    runGenericTest(name, parserHelper, config);
 }
 
 function countActualSymbols(text: string): number {