+++ CHANGELOG.md
@@ -4,6 +4,51 @@ All notable changes to the "OpenEdge ABL Formatter" extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.5.4 (2026-04-24)
+
+- Improved Method Call formatting
+- Improved USING statement formatting
+- Fixed some issues with the Includes
+- Other Improvements [See all](https://github.com/BalticAmadeus/OpenedgeAblFormatter/milestone/22?closed=1)
+
+## 0.5.3 (2026-04-03)
+
+- Added option to disable marketing information
+
+## 0.5.2 (2026-04-02)
+
+- Improved USING statement formatting
+- Added formatting for DELETE statement
+- Added information about upcoming Free Webinar
+
+## 0.5.1 (2026-03-19)
+
+- Implemented new UI experience for changing Formatter settings
+- Improved Array formatting
+- **tree-sitter-abl** version bump and other Improvements [See all](https://github.com/BalticAmadeus/OpenedgeAblFormatter/milestone/20?closed=1)
+
+## 0.5.0 (2026-02-06)
+
+- Returned to synchronous one worker implementation
+- Improved ASSIGN statement formatting
+- Improved UPDATE statement formatting
+- Added configuration for TEMP-TABLE formatting
+- **tree-sitter-abl** version bump and other Improvements [See all](https://github.com/BalticAmadeus/OpenedgeAblFormatter/milestone/19?closed=1)
+
+## 0.4.6 (2025-12-01)
+
+- Improved ASSIGN statement formatting
+- Improved INPUT/OUTPUT statement formatting
+- Fixed issue for selection formatting
+- **tree-sitter-abl** version bump and other Improvements [See all](https://github.com/BalticAmadeus/OpenedgeAblFormatter/milestone/18?closed=1)
+
+## 0.4.5 (2025-11-17)
+
+- Implemented functionality to exclude code from formatting
+- Improved RUN statement formatting
+- Improved MESSAGE statement formatting
+- **tree-sitter-abl** version bump and other Improvements [See all](https://github.com/BalticAmadeus/OpenedgeAblFormatter/milestone/12?closed=1)
+
 ## 0.4.4 (2025-11-07)
 
 - Fixed a startup issue